+++ pastell-core/ActionExecutor.php
@@ -369,6 +369,22 @@ abstract class ActionExecutor
         $this->setLastMessage($message);
     }
 
+    public function changeOrUpdateAction(
+        string $action,
+        string $message,
+        bool $updateDate = false,
+    ): void {
+        $this->objectInstancier->getInstance(ActionChange::class)->addOrUpdateAction(
+            $this->id_d,
+            $this->id_e,
+            $this->id_u,
+            $action,
+            $message,
+            $updateDate
+        );
+        $this->setLastMessage($message);
+    }
+
     public function notify($actionName, $type, $message)
     {
         $this->getNotificationMail()->notify($this->id_e, $this->id_d, $actionName, $type, $message);